+++ backend/clubs/views.py
@@ -1,9 +1,6 @@
 import json
 import os
-import random
 import re
-from collections import OrderedDict
-from urllib.parse import quote
 
 import qrcode
 from django.conf import settings
@@ -21,11 +18,11 @@ from rest_framework import filters, generics, parsers, status, viewsets
 from rest_framework.decorators import action
 from rest_framework.exceptions import PermissionDenied
 from rest_framework.exceptions import ValidationError as DRFValidationError
-from rest_framework.pagination import PageNumberPagination
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
+from clubs.filters import RandomOrderingFilter, RandomPageNumberPagination
 from clubs.mixins import XLSXFormatterMixin
 from clubs.models import (
     Asset,
@@ -89,10 +86,6 @@ from clubs.serializers import (
     YearSerializer,
 )
 from clubs.utils import html_to_text
-
-
-DEFAULT_PAGE_SIZE = 15
-DEFAULT_SEED = 1234
 
 
 def file_upload_endpoint_helper(request, code):
@@ -187,60 +180,6 @@ class ReportViewSet(viewsets.ModelViewSet):
             return Report.objects.none()
 
         return Report.objects.filter(Q(creator=self.request.user) | Q(public=True))
-
-
-class ClubPagination(PageNumberPagination):
-    """
-    Custom pagination for club list view.
-    """
-
-    page_size = DEFAULT_PAGE_SIZE
-    page_size_query_param = "page_size"
-
-    def paginate_queryset(self, queryset, request, view=None):
-        if "random" in request.query_params.get("ordering", "").split(","):
-            rng = random.Random(request.GET.get("seed", DEFAULT_SEED))
-            results = list(queryset)
-            rng.shuffle(results)
-
-            self._random_count = getattr(request, "_original_clubs_count", None)
-            if self._random_count is None:
-                self._random_count = Club.objects.count()
-            else:
-                del request._original_clubs_count
-
-            page = int(request.GET.get("page", 1))
-            page_size = int(request.GET.get("page_size", DEFAULT_PAGE_SIZE))
-
-            if (page - 1) * page_size >= self._random_count:
-                self._random_next_page = None
-            else:
-                new_params = request.GET.dict()
-                new_params["page"] = str(page + 1)
-                self._random_next_page = "{}?{}".format(
-                    request.build_absolute_uri(request.path),
-                    "&".join(["{}={}".format(k, quote(v)) for k, v in new_params.items()]),
-                )
-            return results
-
-        if self.page_query_param not in request.query_params:
-            return None
-
-        return super().paginate_queryset(queryset, request, view)
-
-    def get_paginated_response(self, data):
-        if hasattr(self, "_random_next_page"):
-            return Response(
-                OrderedDict(
-                    [
-                        ("count", self._random_count),
-                        ("next", self._random_next_page),
-                        ("results", data),
-                    ]
-                )
-            )
-
-        return super().get_paginated_response(data)
 
 
 class ClubsSearchFilter(filters.BaseFilterBackend):
@@ -357,7 +296,7 @@ class ClubsSearchFilter(filters.BaseFilterBackend):
         return queryset
 
 
-class ClubsOrderingFilter(filters.OrderingFilter):
+class ClubsOrderingFilter(RandomOrderingFilter):
     """
     Custom ordering filter for club objects.
     """
@@ -382,23 +321,6 @@ class ClubsOrderingFilter(filters.OrderingFilter):
     def filter_queryset(self, request, queryset, view):
         new_queryset = super().filter_queryset(request, queryset, view)
         ordering = request.GET.get("ordering", "").split(",")
-
-        # handle random ordering
-        if "random" in ordering:
-            page = int(request.GET.get("page", 1)) - 1
-            page_size = int(request.GET.get("page_size", DEFAULT_PAGE_SIZE))
-            rng = random.Random(request.GET.get("seed", DEFAULT_SEED))
-
-            all_ids = list(new_queryset.order_by("id").values_list("id", flat=True))
-            rng.shuffle(all_ids)
-
-            start_index = page * page_size
-            end_index = (page + 1) * page_size
-            page_ids = all_ids[start_index:end_index]
-
-            request._original_clubs_count = new_queryset.count()
-
-            return new_queryset.filter(id__in=page_ids)
 
         if "featured" in ordering:
             return queryset.order_by("-rank")
@@ -453,7 +375,7 @@ class ClubViewSet(XLSXFormatterMixin, viewsets.ModelViewSet):
 
     lookup_field = "code"
     http_method_names = ["get", "post", "put", "patch", "delete"]
-    pagination_class = ClubPagination
+    pagination_class = RandomPageNumberPagination
 
     def get_queryset(self):
         queryset = super().get_queryset()
